[style]: renamed package name
