+++ services/auth-service/internal/model/token.go
@@ -1,4 +1,4 @@
-package models
+package model
 
 import (
 	"github.com/google/uuid"
